Unit tests for `SimulatorSettings` decoding.
