+++ Tests/TestArgFileTests/TestArgFileEntryTests.swift
@@ -164,150 +164,6 @@ final class TestArgFileEntryTests: XCTestCase {
         )
     }
     
-    func test___decoding_full_json_without_simulatorKeychainSettings() throws {
-        let json = Data(
-            """
-            {
-                "testsToRun": [
-                    {
-                        "predicateType": "singleTestName",
-                        "testName": "ClassName/testMethod"
-                    }
-                ],
-                "environment": {
-                    "value": "key"
-                },
-                "numberOfRetries": 42,
-                "testDestination": {
-                    "deviceType": "iPhone SE",
-                    "runtime": "11.3"
-                },
-                "testType": "logicTest",
-                "buildArtifacts": {
-                    "appBundle": {
-                        "url": "/appBundle"
-                    },
-                    "runner": {
-                        "url": "/runner"
-                    },
-                    "xcTestBundle": {
-                        "location": {
-                            "url": "/xcTestBundle"
-                        },
-                        "testDiscoveryMode": "runtimeAppTest"
-                    },
-                    "additionalApplicationBundles": ["/additionalApp1","/additionalApp2"]
-                },
-                "testRunnerTool": {
-                    "toolType": "xcodebuild"
-                },
-                "simulatorControlTool": {
-                    "location": "insideUserLibrary",
-                    "tool": {
-                        "toolType": "simctl"
-                    }
-                },
-                "developerDir": {
-                    "kind": "current"
-                },
-                "pluginLocations": [
-                    {
-                        "url":"http://example.com/plugin.zip#sample.emceeplugin"
-                    }
-                ],
-                "scheduleStrategy": "unsplit",
-                "simulatorOperationTimeouts": {
-                    "create": 50,
-                    "boot": 51,
-                    "delete": 52,
-                    "shutdown": 53,
-                    "automaticSimulatorShutdown": 54,
-                    "automaticSimulatorDelete": 55
-                },
-                "simulatorSettings": {
-                    "simulatorLocalizationSettings": {
-                        "localeIdentifier": "ru_US",
-                        "keyboards": [
-                            "ru_RU@sw=Russian;hw=Automatic",
-                            "en_US@sw=QWERTY;hw=Automatic"
-                        ],
-                        "passcodeKeyboards": [
-                            "ru_RU@sw=Russian;hw=Automatic",
-                            "en_US@sw=QWERTY;hw=Automatic"
-                        ],
-                        "languages": [
-                            "ru-US",
-                            "en",
-                            "ru-RU"
-                        ],
-                        "addingEmojiKeybordHandled": true,
-                        "enableKeyboardExpansion": true,
-                        "didShowInternationalInfoAlert": true,
-                        "didShowContinuousPathIntroduction": true
-                    },
-                    "watchdogSettings": {
-                        "bundleIds": [
-                            "sample.app"
-                        ],
-                        "timeout": 42
-                    },
-                },
-                "testTimeoutConfiguration": {
-                    "singleTestMaximumDuration": 42,
-                    "testRunnerMaximumSilenceDuration": 24
-                },
-                "workerCapabilityRequirements": []
-            }
-            """.utf8
-        )
-        
-        let entry = assertDoesNotThrow {
-            try JSONDecoder().decode(TestArgFileEntry.self, from: json)
-        }
-
-        XCTAssertEqual(
-            entry,
-            TestArgFileEntry(
-                buildArtifacts: buildArtifacts(),
-                developerDir: .current,
-                environment: ["value": "key"],
-                numberOfRetries: 42,
-                pluginLocations: [
-                    PluginLocation(.remoteUrl(URL(string: "http://example.com/plugin.zip#sample.emceeplugin")!, [:]))
-                ],
-                scheduleStrategy: .unsplit,
-                simulatorControlTool: SimulatorControlTool(
-                    location: .insideUserLibrary,
-                    tool: .simctl
-                ),
-                simulatorOperationTimeouts: SimulatorOperationTimeouts(
-                    create: 50,
-                    boot: 51,
-                    delete: 52,
-                    shutdown: 53,
-                    automaticSimulatorShutdown: 54,
-                    automaticSimulatorDelete: 55
-                ),
-                simulatorSettings: SimulatorSettings(
-                    simulatorLocalizationSettings: SimulatorLocalizationSettingsFixture().simulatorLocalizationSettings(),
-                    simulatorKeychainSettings: SimulatorKeychainSettings(
-                        rootCerts: []
-                    ),
-                    watchdogSettings: WatchdogSettings(bundleIds: ["sample.app"], timeout: 42)
-                ),
-                testDestination: try TestDestination(deviceType: "iPhone SE", runtime: "11.3"),
-                testRunnerTool: .xcodebuild,
-                testTimeoutConfiguration: TestTimeoutConfiguration(
-                    singleTestMaximumDuration: 42,
-                    testRunnerMaximumSilenceDuration: 24
-                ),
-                testType: .logicTest,
-                testsToRun: [.testName(TestName(className: "ClassName", methodName: "testMethod"))],
-                workerCapabilityRequirements: []
-            )
-        )
-    }
-    
     func test___decoding_short_json() throws {
         let json = Data(
             """
